+++ main.go
@@ -46,7 +46,7 @@ func main() {
 			args.Create.Protection.IncludeAdmins,
 			args.Create.Protection.CanDismiss, args.Create.Protection.CanDismissTeams,
 			args.Create.Protection.CanPush, args.Create.Protection.CanPushTeams,
-			[]string{})
+			args.Create.Protection.RequiredStatusChecks)
 	}
 }